fixup cyclic dependencies
