+++ karapace/schema_type.py
@@ -1,0 +1,8 @@
+from enum import Enum, unique
+
+
+@unique
+class SchemaType(str, Enum):
+    AVRO = "AVRO"
+    JSONSCHEMA = "JSON"
+    PROTOBUF = "PROTOBUF"
